Update to 0.7.0
2003-06-06  Marco Pesenti Gritti  <marco@it.gnome.org>

	* configure.in:

	Update to 0.7.0

	* lib/egg/egg-menu-merge.c: (egg_menu_merge_add_ui_from_string):
	* lib/egg/egg-menu-merge.h:

	Fix crash on startup on alpha

+++ lib/egg/egg-menu-merge.c
@@ -700,7 +700,7 @@ static GMarkupParser ui_parser = {
 
 guint
 egg_menu_merge_add_ui_from_string (EggMenuMerge *self,
-				   const gchar *buffer, guint length,
+				   const gchar *buffer, glong length,
 				   GError **error)
 {
   ParseContext ctx = { 0 };
